+++ modules/internal/src/main/scala/higherkindness/mu/rpc/internal/metrics/MetricsOps.scala
@@ -16,7 +16,6 @@
 
 package higherkindness.mu.rpc.internal.metrics
 import higherkindness.mu.rpc.internal.interceptors.GrpcMethodInfo
-import io.grpc.MethodDescriptor.MethodType._
 import io.grpc.Status
 
 trait MetricsOps[F[_]] {
@@ -39,52 +38,5 @@ trait MetricsOps[F[_]] {
       status: Status,
       elapsed: Long,
       classifier: Option[String]): F[Unit]
-
-}
-
-object MetricsOps {
-
-  def eventDescription(
-      prefix: String,
-      classifier: Option[String],
-      methodInfo: GrpcMethodInfo,
-      eventName: String,
-      status: Option[Status] = None) =
-    classifier
-      .map(c => s"$prefix.$c")
-      .getOrElse(s"$prefix.default") + "." + methodInfoDescription(methodInfo) + s".$eventName" + status
-      .map(s => s".${statusDescription(s)}")
-      .getOrElse("")
-
-  private def methodInfoDescription(methodInfo: GrpcMethodInfo): String =
-    methodInfo.`type` match {
-      case UNARY => s"${methodInfo.serviceName}.${methodInfo.methodName}.unary"
-      case CLIENT_STREAMING =>
-        s"${methodInfo.serviceName}.${methodInfo.methodName}.client-streaming"
-      case SERVER_STREAMING =>
-        s"${methodInfo.serviceName}.${methodInfo.methodName}.server-streaming"
-      case BIDI_STREAMING => s"${methodInfo.serviceName}.${methodInfo.methodName}.bidi-streaming"
-      case UNKNOWN        => s"${methodInfo.serviceName}.${methodInfo.methodName}.unknown"
-    }
-
-  private def statusDescription(status: Status): String = status match {
-    case Status.ABORTED             => "aborted"
-    case Status.ALREADY_EXISTS      => "already-exists"
-    case Status.CANCELLED           => "cancelled"
-    case Status.DATA_LOSS           => "data-loss"
-    case Status.DEADLINE_EXCEEDED   => "deadline-exceeded"
-    case Status.FAILED_PRECONDITION => "failed-precondition"
-    case Status.INTERNAL            => "internal"
-    case Status.INVALID_ARGUMENT    => "invalid-argument"
-    case Status.NOT_FOUND           => "not-found"
-    case Status.OK                  => "ok"
-    case Status.OUT_OF_RANGE        => "out-of-range"
-    case Status.PERMISSION_DENIED   => "permission-denied"
-    case Status.RESOURCE_EXHAUSTED  => "resource-exhausted"
-    case Status.UNAUTHENTICATED     => "unauthenticated"
-    case Status.UNAVAILABLE         => "unavailable"
-    case Status.UNIMPLEMENTED       => "unimplemented"
-    case Status.UNKNOWN             => "unknown"
-  }
 
 }
